[FE] refactor: 회원가입 컴포넌트 분리
