+++ FE/src/styles/auth/Signup.module.scss
@@ -27,49 +27,51 @@ $color-gray-5: var(--color-gray-5);
   margin-bottom: 1.5em;
 }
 
-.signup_box_1 {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-
-  .input_box__button {
-    display: flex;
+.signup-account {
+  &__form {
     width: 100%;
-    gap: 0.3em;
-    align-items: baseline;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
 
-    > :nth-child(2) {
-      min-width: 22%;
+    .input_box__button {
+      display: flex;
+      width: 100%;
+      gap: 0.3em;
+      align-items: baseline;
+
+      > :nth-child(2) {
+        min-width: 22%;
+      }
     }
+  }
+
+  &__warning {
+    margin: 1em 0;
+    height: 2.5em;
+    text-align: center;
+    color: $color-warning;
   }
 }
 
-.signup_warning {
-  margin: 1em 0;
-  height: 2.5em;
-  text-align: center;
-  color: $color-warning;
+.signup-account {
+  &__button {
+    margin-bottom: 1.5em;
+    align-items: center;
+  }
 }
 
-.signup_box_2 {
+.signup_context {
   display: flex;
-  gap: 1.5em;
-  flex-direction: column;
-  align-items: center;
+  gap: 0.5rem;
 
-  > .signup_context {
-    display: flex;
-    gap: 0.5rem;
+  > span:nth-child(1) {
+    color: $color-gray-4;
+  }
 
-    > span:nth-child(1) {
-      color: $color-gray-4;
-    }
-
-    > a {
-      color: $color-gray-5;
-      text-decoration-line: underline;
-    }
+  > a {
+    color: $color-gray-5;
+    text-decoration-line: underline;
   }
 }
